[PP][OPS-10176] set cookie in action
