+++ conf/application.conf
@@ -61,7 +61,6 @@ play.i18n.langs = ["en"]
 play.http.router = prod.Routes
 
 play.filters.enabled += "play.filters.cors.CORSFilter"
-play.filters.enabled += "deviceid.ScpStubDeviceIdFilter"
 
 play.filters.cors {
     allowedOrigins = [
